refactor(store): extract selector models and metadata to own file
