+++ packages/store/src/selectors/selector-models.ts
@@ -1,0 +1,12 @@
+import { SharedSelectorOptions, SelectFromRootState } from '../internal/internals';
+
+export interface CreationMetadata {
+  containerClass: any;
+  selectorName: string;
+  getSelectorOptions?: () => SharedSelectorOptions;
+}
+
+export interface RuntimeSelectorInfo {
+  selectorOptions: SharedSelectorOptions;
+  argumentSelectorFunctions: SelectFromRootState[];
+}
